XXE Test and travis updated with both unit and integration test phase
