+++ src/it/resources/lessons/xxe.properties
@@ -1,0 +1,2 @@
+xxe.lesson.file=briefcase.pulp
+xxe.lesson.solution=Marsellus Wallace's Key: c8c232cd8e3abdfea3fcef24379415a65e00ac197af09e82602e231aa3031879
