+++ src/main/java/servlets/Setup.java
@@ -28,7 +28,6 @@ import org.apache.log4j.Logger;
 import dbProcs.Constants;
 import dbProcs.Database;
 import servlets.module.lesson.XxeLesson;
-import utils.FileSystem;
 import utils.InstallationException;
 import utils.Validate;
